Moved queries to lib instead
