+++ apps/web/components/team/TeamSettings.tsx
@@ -4,8 +4,8 @@ import React, { useRef, useState } from "react";
 import { useLocale } from "@calcom/lib/hooks/useLocale";
 import showToast from "@calcom/lib/notification";
 import { objectKeys } from "@calcom/lib/objectKeys";
+import { TeamWithMembers } from "@calcom/lib/server/queries/teams";
 import { trpc } from "@calcom/trpc/react";
-import { TeamWithMembers } from "@calcom/trpc/server/queries/teams";
 import { Alert } from "@calcom/ui/Alert";
 import Button from "@calcom/ui/Button";
 import { TextField } from "@calcom/ui/form/fields";
